pyproject.toml
[project]
name = "discord-ai-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "abdelazizr", email = "flipaziz@yahoo.com" }
]
dependencies = [
    "discord.py>=2.3.2",
    "openai",
    "python-dotenv",
    "replit>=3.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/discord_ai_bot"]